move loadingdots component to fiftyone/components
