+++ app/packages/core/src/components/Actions/ActionsRow.tsx
@@ -35,7 +35,7 @@ import { useEventHandler, useOutsideClick, useSetView } from "@fiftyone/state";
 import Similar from "./Similar";
 import { useTheme } from "@fiftyone/components";
 import * as fos from "@fiftyone/state";
-import LoadingDots from "../Common/LoadingDots";
+import LoadingDots from "../../../../components/src/components/Loading/LoadingDots";
 
 const Loading = () => {
   const theme = useTheme();
